refact(errors): reorganize files
