+++ src/utils/errors/http-error.js
@@ -1,0 +1,9 @@
+class HttpError extends Error {
+  constructor(response) {
+    super(response.statusText);
+    this.code     = response.status;
+    this.response = response;
+  }
+}
+
+export default HttpError;
