+++ src/utils/errors/index.ts
@@ -1,1 +1,4 @@
-export  * from './errors.js';
+export { default as HttpError }       from './http-error.js';
+export { default as ValidationError } from './validation-error.js';
+
+export * from './errors.js';
