+++ src/utils/errors/validation-error.js
@@ -1,11 +1,3 @@
-class HttpError extends Error {
-  constructor(response) {
-    super(response.statusText);
-    this.code     = response.status;
-    this.response = response;
-  }
-}
-
 class ValidationError extends Error {
   constructor(errors) {
     super('validation error');
@@ -18,4 +10,4 @@ class ValidationError extends Error {
   }
 }
 
-export { HttpError, ValidationError };
+export default ValidationError;
